third_party/WebKit/LayoutTests/fast/events/scale-and-scroll-div-expected.txt

This tests that a div scrolled by gesture touch while the page is scaled still scrolls at the rate of the touch

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
Scrolling unscaled div
PASS scrollBox.scrollTop is expectedScrollDistance
PASS scrollBox.scrollTop is expectedScrollDistance
Scrolling 1.5 scaled div
PASS scrollBox.scrollTop is expectedScrollDistance
PASS scrollBox.scrollTop is expectedScrollDistance
Scrolling 1.63 scaled div
PASS scrollBox.scrollTop is expectedScrollDistance
PASS scrollBox.scrollTop is expectedScrollDistance
Scrolling 2.0 scaled div
PASS scrollBox.scrollTop is expectedScrollDistance
PASS scrollBox.scrollTop is expectedScrollDistance
